+++ src/core/bau_inputs.jl
@@ -322,10 +322,10 @@ function bau_outage_check(critical_loads_kw::AbstractArray, pv_kw_series::Abstra
         if length(pv_kw_series) == 0
             pv_kw_series = zeros(length(critical_loads_kw))
         end
-        fuel_slope_gal_per_kwhe, fuel_intercept_gal_per_hr = generator_fuel_slope_and_intercept(
+        fuel_slope_gal_per_kwhe, fuel_intercept_gal_per_hr = fuel_slope_and_intercept(
             electric_efficiency_full_load=gen.electric_efficiency_full_load, 
             electric_efficiency_half_load=gen.electric_efficiency_half_load,
-            fuel_higher_heating_value_kwh_per_gal=gen.fuel_higher_heating_value_kwh_per_gal
+            fuel_higher_heating_value_kwh_per_unit=gen.fuel_higher_heating_value_kwh_per_gal
         )
             for (i, (load, pv)) in enumerate(zip(critical_loads_kw, pv_kw_series))
             unmet = load - pv